Added ability to inject AppState at first config time
