+++ .github/copilot-instructions.md
@@ -1,0 +1,1 @@
+Always use System.text.json for working with JSON markup
