+++ .github/workflows/dotnet-build.yml
@@ -50,7 +50,6 @@ jobs:
 
     - name: Set badge color
       shell: bash
-      if: always()
       run: |
        case ${{ fromJSON( steps.test-results.outputs.json ).conclusion }} in
          success)
@@ -64,7 +63,6 @@ jobs:
            ;;
        esac
     - name: Create badge
-      if: always()
       uses: emibcn/badge-action@808173dd03e2f30c980d03ee49e181626088eee8
       with:
        label: Unit Tests
